change type for baseline ranx

+++ redisvl/utils/optimize/cache.py
@@ -12,19 +12,19 @@ from redisvl.utils.optimize.utils import NULL_RESPONSE_KEY, _format_qrels
 
 def _generate_run_cache(test_data: List[LabeledData], threshold: float) -> Run:
     """Format observed data for evaluation with ranx"""
-    run_dict: Dict[str, Dict[str, int]] = {}
+    run_dict: Dict[str, Dict[str, float]] = {}
 
     for td in test_data:
         run_dict[td.id] = {}
         for res in td.response:
             if float(res["vector_distance"]) < threshold:
                 # value of 1 is irrelevant checks only on match for f1
-                run_dict[td.id][res["id"]] = 1
+                run_dict[td.id][res["id"]] = 1.0
 
         if not run_dict[td.id]:
             # ranx is a little odd in that if there are no matches it errors
             # if however there are no keys that match you get the correct score
-            run_dict[td.id][NULL_RESPONSE_KEY] = 1
+            run_dict[td.id][NULL_RESPONSE_KEY] = 1.0
 
     return Run(run_dict)
